Split Parakeet audio ownership support
